Generate sources when required, not when building utilities
Instead of running ./generateConstants and ./ttf2RasterFonts in the
rules that build them, run them when the sources are required.

To help with parallel makes, use timestamp files .stamp-constantPointers
and .stamp-rasterFontsData to mark when the sources have been generated.
This is especially important for .stamp-constantPointers since it marks
that two different source files have been generated.