Resolve "STO+ crashes the simulator!"
The result of an arithmetic operation doesn't usually go in the X
register for the storage arithmetic operations, so various assumptions
about the initial type of the result register matching the final type of
the result register are incorrect, leading to crashes.  Currently, only
the addition, subtraction, multiplication and division operations are
affected.  Fix it by reallocating the result register where necessary.