Change .gitignore to ignore built executables in top-level directory
The './' prefix on the .gitignore rules to ignore the executables in the
top-level directory such as 'wp43s' doesn't seem to work.  The main
reason for adding the './' prefic was to prevent the subdirectories of
'src/' such as 'src/wp43s' being ignored, and that does work.  Fix it by
changing the prefix to '/' to anchor the rules to the top-level
directory only.