Fix test for exceeding MAX_FREE_BLOCKS in wp43sFree()
In `wp43sFree()`, correct the test for whether the maximum number of
free blocks is being exceeded.  The current test allows
`numberOfFreeBlocks` to be incremented beyond `MAX_FREE_BLOCKS` because
it is testing the wrong variable.

Also, move the test so it is done just before it moves information about
free blocks around to make room for the new free block.